Angles should be in radians, except for UI / IO purposes
Angles internally (inputs and outputs to functions, stored in
data members etc) should ALWAYS be in radians, and we should ideally
use the half-open range [-PI .. +PI) for these values.
Here you are using 0 .. 360 in degrees, which is not ideal. I can't
change this without breaking your horizontal / vertical classification
code, so I leave this comment in as a design note.

Also, this code not only returns degrees, but also rounds that off to
2 decimals which is not enough precision for most of our purposes. In
general API like this should not round off the returned value at all -
that should be done only during comparison for equality, or during
display to the user.