Don't use Math.Pow for small integral powers
Math.Pow is expensive: it computes the natural logarithm of the number,
multiplies that by the power, and finally computes e^x of the result.
Both are expensive operations, and for just squaring or cubing a
double, it is must faster to just use multiplications.

Also, this code is rounding off the distance to 2 decimals before returning,
which is not good behavior for a function - what if you need more precision
in some situation. The caller should do this rounding as they need, so I
removed this rounding. It is possible this breaks your code, so check the
places where you are calling this and re-introducing some rounding as you need.
10^-2 is too aggressive, however. I would recommend a consistent 1e-6 everywhere
for comparison.